refactor: extract ball carry stats accumulator
